% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart2tidy_function.R
\name{rpart2tidy}
\alias{rpart2tidy}
\title{Turns rpart data into tibble}
\usage{
rpart2tidy(rpart_object, text_size = 4, node_spacing = 0.5)
}
\arguments{
\item{rpart_object}{the name of the rpart object}

\item{text_size}{number, defaults to 4}

\item{node_spacing}{the vertical distance between nodes for purity and nodes for splits values. Defaults to 0.5}
}
\value{
a list of three dataframes (or tibbles)
\enumerate{
   \item nodes.tbl has data for the main nodes
   \item linesegments.tbl has the data for connecting lines
   \item YES_NO.tbl has xy coordinates for small yes-no branching nodes
}
}
\description{
This function is an alternative to rpart.plot and other graphing options for decision trees objects generated by rpart
}
